+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Version 0.3.10 - 2024-xx-xx
 
 ### Added
+
+* server-side code-coverage
+
 ### Changed
 ### Removed
 ### Fixed